bladewolf spin attack
created new attack with explicit triggers to aim for a player and spin towards them *only if* player is above the boss by a certain distance; this is an attempt to consider that players could otherwise cheese the boss because bladewolf only has one ranged attack...

again, this is the "unstable and untested" branch, so functionality of this attack is not yet determined. will test this properly when I can.